Move truncate to worker folder
