+++ worker/lib/truncate.js
